backend/pyproject.toml
[project]
name = "signal-hunter"
version = "0.1.0"
description = "Decision intelligence terminal for markets"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0,<1.0",
    "uvicorn[standard]>=0.32.0,<1.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.36,<3.0",
    "asyncpg>=0.30.0,<1.0",
    "alembic>=1.14.0,<2.0",

    # Supabase client (for Realtime + REST)
    "supabase>=2.11.0,<3.0",

    # Redis
    "upstash-redis>=1.1.0,<2.0",

    # AI
    "anthropic>=0.42.0,<1.0",

    # Validation & settings
    "pydantic>=2.10.0,<3.0",
    "pydantic-settings>=2.6.0,<3.0",
    "python-dotenv>=1.0.0,<2.0",

    # HTTP client (for sensors)
    "httpx>=0.28.0,<1.0",

    # Scheduling
    "apscheduler>=4.0.0a5",

    # Data
    "python-dateutil>=2.9.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9.0",
    "pytest-asyncio>=0.24.0,<1.0",
    "pytest-cov>=6.0.0,<7.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-python-dateutil>=2.9.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "T20",  # flake8-print
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific
]
ignore = [
    "B008",   # Depends() in function defaults — FastAPI pattern
    "RUF002", # Ambiguous unicode in docstrings (multiplication sign)
]

[tool.ruff.lint.per-file-ignores]
"alembic/**/*.py" = ["E402"]  # Alembic env.py requires late imports

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["supabase.*", "upstash_redis.*", "apscheduler.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
